Refactoring the highlighter and commenting
