+++ components/Markdown/CodeHighlight.tsx
@@ -5,25 +5,31 @@ import { PreviewObject } from 'figma-exporter/src/types';
 import CopyCode from 'components/CopyCode';
 SyntaxHighlighter.registerLanguage('html', html);
 
+/**
+ * Highlight code for preview elements
+ * @param param0
+ * @returns ReactElement
+ */
 export const CodeHighlight: React.FC<{ data: PreviewObject | undefined }> = ({ data }) => {
-    if (data) {
-      return (
-        <div className="c-code-block">
-          <SyntaxHighlighter
+  if (data) {
+    return (
+      <div className="c-code-block">
+        <SyntaxHighlighter
           // @ts-ignore
           style={oneLight}
-          language='html'
+          language="html"
           PreTag="div"
           showLineNumbers={true}
           wrapLines={true}
           useInlineStyles={true}
         >
-            {data.code}
-          </SyntaxHighlighter>
-          <CopyCode code={data.code} />
-        </div>
-      );
-    } else {
-      return <></>;
-    }
-  };
+          {data.code}
+        </SyntaxHighlighter>
+        <CopyCode code={data.code} />
+      </div>
+    );
+  } else {
+    return <></>;
+  }
+};
+
