+++ pages/[level1]/[level2]/index.tsx
@@ -5,7 +5,7 @@ import Head from 'next/head';
 import Header from 'components/Header';
 import CustomNav from 'components/SideNav/Custom';
 import { ReactMarkdown } from 'react-markdown/lib/react-markdown';
-import { MarkdownComponents } from 'components/util/MarkdownComponents';
+import { MarkdownComponents } from 'components/Markdown/MarkdownComponents';
 import rehypeRaw from 'rehype-raw';
 
 export interface SubPageType {
