+++ pages/assets/index.tsx
@@ -9,7 +9,7 @@ import { DocumentationProps, fetchDocPageMarkdown } from 'components/util';
 import Header from 'components/Header';
 import ReactMarkdown from 'react-markdown';
 import CustomNav from 'components/SideNav/Custom';
-import { MarkdownComponents } from 'components/util/MarkdownComponents';
+import { MarkdownComponents } from 'components/Markdown/MarkdownComponents';
 import rehypeRaw from 'rehype-raw';
 
 const config = getConfig();
