+++ pages/components/index.tsx
@@ -7,7 +7,7 @@ import { DocumentationProps, fetchDocPageMarkdown, fetchDocPageMetadataAndConten
 import Header from 'components/Header';
 import ReactMarkdown from 'react-markdown';
 import CustomNav from 'components/SideNav/Custom';
-import { MarkdownComponents } from 'components/util/MarkdownComponents';
+import { MarkdownComponents } from 'components/Markdown/MarkdownComponents';
 import rehypeRaw from 'rehype-raw';
 
 enum AvailableComponentPageComponents {
