+++ pages/foundations/logo.tsx
@@ -11,7 +11,7 @@ import { DocumentationProps, fetchDocPageMarkdown } from 'components/util';
 import Header from 'components/Header';
 import CustomNav from 'components/SideNav/Custom';
 import { ReactMarkdown } from 'react-markdown/lib/react-markdown';
-import { MarkdownComponents } from 'components/util/MarkdownComponents';
+import { MarkdownComponents } from 'components/Markdown/MarkdownComponents';
 import rehypeRaw from 'rehype-raw';
 
 const config = getConfig();
